Initial test of using spicommon_bus_initialize_io() for WS2801
There is a known bug with use of DMA in ESP32 that resets entire DMA channel for both SPI2 and SPI3 when either one is reset.  This means calling spi_bus_free() for WS2801 in order to switch pins when using multiple LED strands crashes the other SPI bus.  This is problematic is other bus is being used by sketch for something else, such as Ethernet connectivity.

Instead of calling spi_bus_free() this new (test) logic calls an internal IDF functions that re-assigns pins to SPI bus instead of re-initializing the bus, which means we can avoid calling spi_bus_free().